refactor: extract CheckResultsGrid, remove shim modules; add codebase review
Two low-risk cleanups surfaced during a full codebase review (see
`docs/reviews/2026-07-11-codebase-review.md`):

Frontend — extract CheckResultsGrid from SkillDetailPage.tsx
- Move the `CheckResultsGrid` component + its `formatCheckName` helper out
  of the 785-line `pages/SkillDetailPage.tsx` into
  `components/CheckResultsGrid.tsx` + `components/formatCheckName.ts`, with
  a matching `CheckResultsGrid.module.css` for the ~110 lines of CSS that
  belonged to it.
- The existing test file `pages/CheckResultsGrid.test.tsx` referenced a
  source path that didn't exist (the component was defined inside
  `SkillDetailPage.tsx`). Renaming the test to
  `components/CheckResultsGrid.test.tsx` makes the filename match the
  subject and removes the orphan.
- Deleted `pages/auditUtils.ts` (single-caller util now co-located with
  the component that uses it).
- `formatCheckName` lives in a sibling file rather than co-exported to
  keep the component file react-refresh clean.

Client — drop two shim modules that only re-exported from `dhub_core`
- Deleted `client/src/dhub/models.py` (14 lines, pure re-export of
  `dhub_core.models`).
- Deleted `client/src/dhub/core/manifest.py` (12 lines, pure re-export of
  `dhub_core.manifest`).
- Updated 7 call sites (4 source, 3 tests) to import from `dhub_core.*`
  directly, eliminating "two ways to import the same symbol" in the
  client. `_SKILL_NAME_PATTERN` (used by one test) is imported from its
  home in `dhub_core.validation` instead of via the manifest shim.

`client/src/dhub/core/validation.py` is kept because it still owns
`parse_skill_ref`, which is client-only.

Verification:
- `make test-shared` — 98 passed
- `make test-client` — 291 passed, 29 skipped (baseline unchanged)
- `frontend` vitest — 82 passed (baseline unchanged), the extracted
  CheckResultsGrid tests now live under `components/`
- `make lint-frontend` — clean (one pre-existing warning at
  SkillDetailPage.tsx:100 unrelated to this change)

Docs
- Adds `docs/reviews/2026-07-11-codebase-review.md` — a principal-engineer
  review of the whole repo (architecture, code health, security, testing,
  observability, DX) with a top-10 leverage list and a 5-phase migration
  plan. This PR delivers phase 1.
