+++ frontend/src/components/formatCheckName.ts
