+++ frontend/src/pages/SkillDetailPage.module.css
@@ -500,119 +500,6 @@ a.metaItem:hover {
   color: var(--text-muted);
 }
 
-.auditChecks {
-  margin-top: 4px;
-}
-
-.auditCheckTitle {
-  font-family: 'Rajdhani', sans-serif;
-  font-size: var(--text-xxs);
-  font-weight: 600;
-  letter-spacing: 2px;
-  text-transform: uppercase;
-  color: var(--text-muted);
-  margin-bottom: 6px;
-}
-
-.checkGrid {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 8px;
-}
-
-.checkCard {
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-  padding: 10px 12px;
-  border-radius: 4px;
-  background: var(--bg-dark);
-  border: 1px solid var(--border-color);
-  overflow: hidden;
-  cursor: pointer;
-}
-
-.checkHeader {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-}
-
-.checkIcon {
-  flex-shrink: 0;
-}
-
-.checkName {
-  font-family: 'Orbitron', sans-serif;
-  font-size: var(--text-xxs);
-  font-weight: 700;
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  white-space: nowrap;
-  padding: 2px 8px;
-  border-radius: 4px;
-  border: 1px solid;
-}
-
-.checkMessage {
-  font-family: 'Rajdhani', sans-serif;
-  font-size: var(--text-xs);
-  color: var(--text-secondary);
-  line-height: 1.3;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-}
-
-.checkCard:hover .checkMessage,
-.checkMessageExpanded {
-  -webkit-line-clamp: unset;
-}
-
-/* Severity variants */
-.severityPass {
-  border-color: rgba(57, 255, 20, 0.15);
-}
-
-.severityPass .checkIcon {
-  color: var(--neon-green);
-}
-
-.severityPass .checkName {
-  color: var(--neon-green);
-  border-color: rgba(57, 255, 20, 0.3);
-  background: rgba(57, 255, 20, 0.08);
-}
-
-.severityFail {
-  border-color: rgba(255, 45, 149, 0.15);
-}
-
-.severityFail .checkIcon {
-  color: var(--neon-pink);
-}
-
-.severityFail .checkName {
-  color: var(--neon-pink);
-  border-color: rgba(255, 45, 149, 0.3);
-  background: rgba(255, 45, 149, 0.08);
-}
-
-.severityWarn {
-  border-color: rgba(255, 230, 0, 0.15);
-}
-
-.severityWarn .checkIcon {
-  color: var(--neon-yellow);
-}
-
-.severityWarn .checkName {
-  color: var(--neon-yellow);
-  border-color: rgba(255, 230, 0, 0.3);
-  background: rgba(255, 230, 0, 0.08);
-}
-
 .auditQuarantine {
   font-family: 'Share Tech Mono', monospace;
   font-size: var(--text-xs);
